requirements.txt
absl-py==2.0.0
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
asgiref==3.7.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
attrs==23.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
backports.zoneinfo==0.2.1
brotlipy==0.7.0
cachetools==5.3.2
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contourpy==1.1.1
crcmod==1.7
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Django==4.2.7
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock @ file:///croot/filelock_1672387128942/work
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna @ file:///croot/idna_1666125576474/work
image==1.5.33
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1698244021190/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
jmespath==0.10.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1669775081186/work
keras==2.13.1
kiwisolver==1.4.5
labelme2coco==0.2.4
libclang==16.0.6
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv==2.1.0
# Editable install with no version control (mmdet==3.2.0)
-e /home/wzhmiasanmia/miniconda3/envs/openmmlab/lib/python3.8/site-packages
mmengine==0.9.0
model-index==0.1.11
mpmath @ file:///croot/mpmath_1690848262763/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.7.0.72
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.27
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///croot/pillow_1696580024257/work
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.25.0
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.19.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2023.3.post1
PyYAML @ file:///croot/pyyaml_1670514731622/work
pyzmq @ file:///croot/pyzmq_1686601365461/work
referencing==0.31.1
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.13.2
rsa==4.9
sahi==0.11.15
scipy==1.10.1
shapely==2.0.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse==0.4.4
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy @ file:///croot/sympy_1668202399572/work
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminaltables==3.1.10
tomli==2.0.1
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.65.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1698671135544/work
triton==2.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1686163155763/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1698744702785/work
Werkzeug==3.0.1
wrapt==1.16.0
yapf==0.40.2
zipp==3.17.0
